+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # package:sentry changelog
 
+## 0.0.3
+
+- environment attributes
+- auto-generate event_id and timestamp for events
+
 ## 0.0.2
 
 - parse and report stack traces